Rename wfcroom.py because it's not a room like multiroom
