+++ minigrid/envs/wfc/__init__.py
@@ -1,4 +1,4 @@
 from __future__ import annotations
 
 from minigrid.envs.wfc.config import WFC_PRESETS, WFC_PRESETS_SLOW, WFCConfig
-from minigrid.envs.wfc.wfcroom import WFCEnv
+from minigrid.envs.wfc.wfcenv import WFCEnv
